🔀 chore(package.json): add dayjs package as a dependency to handle date and time operations
🔄 refactor(page.tsx): replace createFmcScrambleToNormalScramble function with makeTodaysScramble function to generate today's scramble and display it on the page

🔄 refactor(makeScramble.ts): remove createFmcScrambleToNormalScramble function and replace it with makeTodaysScramble function to generate today's scramble using Scrambo library and cube notation normalizer

✨ feat(scramble.ts): add makeTodaysScramble function to generate today's scramble using Scrambo library, cube notation normalizer, and Prisma ORM. If a scramble for today already exists in the database, retrieve it. Otherwise, create a new scramble and store it in the database.
