+++ package.json
@@ -23,6 +23,7 @@
     "@types/react": "18.2.18",
     "@types/react-dom": "18.2.7",
     "cube-notation-normalizer": "^1.0.0",
+    "dayjs": "^1.11.9",
     "eslint": "8.46.0",
     "eslint-config-next": "13.4.12",
     "next": "13.4.12",
