+++ src/app/page.tsx
@@ -1,17 +1,19 @@
 import { LoginButton, LogoutButton } from '@/components/molecules/LoginButton'
 import { authOptions } from '@/resources/options'
-import { createFmcScrambleToNormalScramble } from '@/services/makeScramble'
+import { makeTodaysScramble } from '@/services/scramble'
 import { getServerSession } from 'next-auth'
 
 export default async function Home() {
   const session = await getServerSession(authOptions)
   const user = session?.user
 
+  const todayScramble = await makeTodaysScramble()
+
   return (
     <div>
       <p>hello {user ? user?.name : 'world'}</p>
       {user ? <LogoutButton /> : <LoginButton />}
-      <p>Scramble: {createFmcScrambleToNormalScramble()}</p>
+      <p>Todays Scramble: {todayScramble.scramble}</p>
     </div>
   )
 }
